# docs/api-docs.yaml
openapi: 3.1.0
info:
  title: Bank API
  version: "1.0"
servers:
- url: http://localhost:8080/bank
  description: Generated server url
security:
- bearerAuth: []
paths:
  /public/user-requests:
    post:
      tags:
      - user-request-public-controller
      operationId: createUserRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestResponse"
  /public/payment/payments/self:
    post:
      tags:
      - payment-public-controller
      operationId: selfPayment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelfPaymentRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: No Content
  /public/card-requests:
    post:
      tags:
      - card-request-public-controller
      operationId: createCardRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestResponse"
  /auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /admin/user-requests/{requestId}/reject:
    post:
      tags:
      - user-request-admin-controller
      operationId: rejectUserRequest
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: No Content
  /admin/user-requests/{requestId}/approve:
    post:
      tags:
      - user-request-admin-controller
      operationId: approveUserRequest
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /admin/card-requests/{requestId}/reject:
    post:
      tags:
      - card-request-admin-controller
      operationId: rejectUserRequest_1
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: No Content
  /admin/card-requests/{requestId}/approve:
    post:
      tags:
      - card-request-admin-controller
      operationId: approveUserRequest_1
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
  /admin/cards/{cardId}/block:
    patch:
      tags:
      - card-admin-controller
      operationId: blockCardByCardId
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
  /admin/cards/by-user/{userId}/block:
    patch:
      tags:
      - card-admin-controller
      operationId: blockAllUsersCards
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
  /public/cards:
    get:
      tags:
      - card-public-controller
      operationId: getMyCards
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
  /public/cards/{cardId}:
    get:
      tags:
      - card-public-controller
      operationId: getMyCardById
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
  /admin/users:
    get:
      tags:
      - user-admin-controller
      operationId: getAllUsers
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
  /admin/users/{userId}:
    get:
      tags:
      - user-admin-controller
      operationId: getUserById
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /admin/user-requests:
    get:
      tags:
      - user-request-admin-controller
      operationId: getAllResponses
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestResponse"
  /admin/user-requests/{requestId}:
    get:
      tags:
      - user-request-admin-controller
      operationId: getResponseById
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestResponse"
  /admin/cards:
    get:
      tags:
      - card-admin-controller
      operationId: getAllCards
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CardResponse"
  /admin/cards/{cardId}:
    get:
      tags:
      - card-admin-controller
      operationId: getCardById
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
  /admin/card-requests:
    get:
      tags:
      - card-request-admin-controller
      operationId: getAllRequests
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestResponse"
  /admin/card-requests/{requestId}:
    get:
      tags:
      - card-request-admin-controller
      operationId: getRequestById
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RequestResponse"
components:
  schemas:
    UserCreateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 250
          minLength: 2
        surname:
          type: string
          maxLength: 250
          minLength: 2
        email:
          type: string
          format: email
          minLength: 1
        age:
          type: integer
          format: int32
          maximum: 120
          minimum: 18
        password:
          type: string
          maxLength: 255
          minLength: 6
      required:
      - age
      - email
      - name
      - password
      - surname
    RequestResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
    SelfPaymentRequest:
      type: object
      properties:
        senderCardId:
          type: integer
          format: int64
        receiverCardId:
          type: integer
          format: int64
        transferAmount:
          type: number
      required:
      - receiverCardId
      - senderCardId
      - transferAmount
    CardCreateRequest:
      type: object
      properties:
        userId:
          type: integer
          format: int64
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - email
      - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        age:
          type: integer
          format: int32
        role:
          type: string
          enum:
          - USER
          - ADMIN
    CardResponse:
      type: object
      properties:
        maskedCardNumber:
          type: string
        expirationDate:
          type: string
          format: date
        balance:
          type: number
        ownerName:
          type: string
        ownerSurname:
          type: string
        cardStatus:
          type: string
          enum:
          - ACTIVE
          - EXPIRED
          - BLOCKED
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT