# src/main/java/com/example/bankcards/dto/request/CardCreateRequest.java
package com.example.bankcards.dto.request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;

@Data
@Builder
@AllArgsConstructor
public class CardCreateRequest {
    private Long userId;
}